sa_family_t is a uint8_t on Darwin but unsigned short on Linux
On Darwin we have:
```
typedef __uint8_t               sa_family_t;

struct sockaddr_in {
        __uint8_t       sin_len;
        sa_family_t     sin_family;
        in_port_t       sin_port;
        struct  in_addr sin_addr;
        char            sin_zero[8];
};

> #define AF_UNIX         1               /* local to host (pipes) */
> #define AF_LOCAL        AF_UNIX         /* backward compatibility */
> #define AF_INET         2               /* internetwork: UDP, TCP, etc. */
> #define AF_INET6        30              /* IPv6 */

```

On Linux we have:
```
typedef unsigned short sa_family_t;

struct sockaddr_in {
	sa_family_t sin_family;
	in_port_t sin_port;
	struct in_addr sin_addr;
	uint8_t sin_zero[8];
};

> #define PF_LOCAL        1
> #define PF_UNIX         PF_LOCAL
> #define AF_LOCAL        PF_LOCAL
> #define PF_INET         2
> #define PF_INET6        10
> #define AF_INET         PF_INET
> #define AF_INET6        PF_INET6
```

Since the relevant constants are all small, we can use a uint8_t on
both platforms.

+++ src/c/luv_c_type_descriptions.ml
@@ -316,7 +316,7 @@ struct
 
     type storage = [ `Sockaddr_storage ] structure
     let storage : storage typ = structure "sockaddr_storage"
-    let family = field storage "ss_family" short
+    let family = field storage "ss_family" int8_t
     let () = seal storage
   end
